Reorganized files and folders structure
