+++ cypress.json
@@ -1,6 +1,5 @@
 {
   "baseUrl": "http://localhost:3000",
-  "defaultCommandTimeout": 2000,
   "video": true,
   "viewportHeight": 1920,
   "viewportWidth": 1080,
