+++ src/components/Comment/Comment.scss
@@ -1,0 +1,3 @@
+.message-body {
+  white-space: pre-line;
+}
