--- src/components/Loader/index.tsx
@@ -1,1 +1,0 @@
-export * from './Loader';
