+++ src/components/Notification/Notification.tsx
@@ -1,0 +1,16 @@
+export const Notification: React.FC = () => {
+  return (
+    <>
+      <div
+        className="notification is-danger"
+        data-cy="PostsLoadingError"
+      >
+        Something went wrong!
+      </div>
+
+      <div className="notification is-warning" data-cy="NoPostsYet">
+        No posts yet
+      </div>
+    </>
+  );
+};
