+++ src/components/Sidebar/Sidebar.scss
@@ -17,7 +17,3 @@
     }
   }
 }
-
-.message-body {
-  white-space: pre-line;
-}
