--- src/types/Comment.ts
@@ -1,9 +1,0 @@
-export interface Comment {
-  id: number;
-  postId: number;
-  name: string;
-  email: string;
-  body: string;
-}
-
-export type CommentData = Pick<Comment, 'name' | 'email' | 'body'>;
