+++ src/types/Post.interface.ts
@@ -1,4 +1,4 @@
-export interface Post {
+export interface IPost {
   id: number;
   userId: number;
   title: string;
