+++ src/types/User.interface.ts
@@ -1,4 +1,4 @@
-export interface User {
+export interface IUser {
   id: number;
   name: string;
   email: string;
